@@ -5,14 +5,13 @@ author:
 homepage:
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.17.0 <2.0.0"
 
 dependencies:
 
-  provider: 5.0.0
-  avatar_glow: 2.0.1
-  auto_size_text: 2.1.0
+  avatar_glow: ^2.0.1
+  auto_size_text: ^3.0.0-nullsafety.0
 
   harcapp_core:
     git:
